refactor: split SettingsPage.tsx into Settings/ section components
SettingsPage.tsx: 2095 → 278 lines. Extracted 15 new files:
- useSettingsPage.ts — custom hook for all state, effects, handlers
- SettingsShared.tsx — AmberToggle, SectionCard, SectionTitle helpers
- ProfileSection, DisplaySection, UnitsSection, DefaultsSection, MapSection
- NotificationsSection, PrivacySection, BackupSection, ParserSection
- AutoUpdateSection, EnrichmentSection, ApiKeysSection
- DeveloperSection, AdminSection, AboutSection
- PasswordModal, DeveloperConfirmDialog
