+++ frontend/src/components/Settings/AdminSection.tsx
@@ -1,0 +1,41 @@
+import { Link } from "react-router-dom";
+import { SectionCard, SectionTitle } from "./SettingsShared";
+import { useTranslation } from "../../hooks/useTranslation";
+
+export default function AdminSection(): JSX.Element {
+  const { t } = useTranslation(["settings"]);
+
+  return (
+    <SectionCard>
+      <SectionTitle
+        title={t("settings:admin.title")}
+        description={t("settings:admin.description")}
+      />
+      <Link
+        to="/admin"
+        className="inline-flex items-center px-4 py-2 font-medium rounded-lg transition-colors"
+        style={{
+          background: "var(--accent)",
+          color: "#0d1117",
+          boxShadow: "0 0 16px rgba(232,160,69,0.25)",
+        }}
+      >
+        <svg className="w-5 h-5 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z"
+          />
+          <path
+            strokeLinecap="round"
+            strokeLinejoin="round"
+            strokeWidth={2}
+            d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"
+          />
+        </svg>
+        {t("settings:admin.openAdminPanel")}
+      </Link>
+    </SectionCard>
+  );
+}
